bootstrap also works if bundle is used from a parent project

+++ Tests/bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 use Doctrine\Common\Annotations\AnnotationRegistry;
 
-if (!is_file($loaderFile = __DIR__.'/../vendor/autoload.php')) {
+if (!is_file($loaderFile = __DIR__.'/../vendor/autoload.php') && !is_file($loaderFile = __DIR__.'/../../../../../autoload.php')) {
     throw new \LogicException('Could not find autoload.php in vendor/. Did you run "composer install --dev"?');
 }
